+++ apps/journeys-admin/src/components/VisitorInfo/JourneyList/EventsCard/CompactEvent/CompactEvent.tsx
