+++ apps/journeys-admin/src/components/VisitorInfo/JourneyList/EventsCard/CompactEvent/index.ts
